Move cythonized client into client_base
- Renames client_callbacks to client_base and includes
  the base part of the client there. This allows us
  to add cythonize types for the connection
