--- aioesphomeapi/client_callbacks.pxd
@@ -1,12 +1,0 @@
-
-cdef dict SUBSCRIBE_STATES_RESPONSE_TYPES
-
-cdef bint TYPE_CHECKING
-
-cdef object CameraImageResponse, CameraState
-
-cdef object HomeassistantServiceCall
-
-cdef object BluetoothLEAdvertisement
-
-cdef object BluetoothDeviceConnectionResponse
